+++ src/controllers/accounts/index.ts
@@ -1,0 +1,3 @@
+export { default as AccountsBalanceInfo } from './AccountsBalanceInfoController';
+export { default as AccountsStakingInfo } from './AccountsStakingInfoController';
+export { default as AccountsVestingInfo } from './AccountsVestingInfoController';
