+++ src/controllers/blocks/index.ts
@@ -1,0 +1,1 @@
+export { default as Blocks } from './BlocksController';
