+++ src/controllers/claims/index.ts
@@ -1,0 +1,1 @@
+export { default as Claims } from './ClaimsController';
