+++ src/controllers/metadata/MetadataController.ts
@@ -2,8 +2,8 @@ import { ApiPromise } from '@polkadot/api';
 import { RequestHandler } from 'express';
 import { INumberParam } from 'src/types/request_types';
 
-import ApiHandler from '../ApiHandler';
-import AbstractController from './AbstractController';
+import { MetadataService } from '../../services';
+import AbstractController from '../AbstractController';
 
 /**
  * GET the chain's metadata.
@@ -19,11 +19,11 @@ import AbstractController from './AbstractController';
  * - FRAME Support: https://crates.parity.io/frame_support/metadata/index.html
  * - Knowledge Base: https://substrate.dev/docs/en/knowledgebase/runtime/metadata
  */
-export default class MetadataController extends AbstractController {
-	handler: ApiHandler;
+export default class MetadataController extends AbstractController<
+	MetadataService
+> {
 	constructor(api: ApiPromise) {
-		super(api, '/metadata');
-		this.handler = new ApiHandler(api);
+		super(api, '/metadata', new MetadataService(api));
 		this.initRoutes();
 	}
 
@@ -46,7 +46,7 @@ export default class MetadataController extends AbstractController {
 
 		MetadataController.sanitizedSend(
 			res,
-			await this.handler.fetchMetadata(hash)
+			await this.service.fetchMetadata(hash)
 		);
 	};
 
@@ -65,7 +65,7 @@ export default class MetadataController extends AbstractController {
 
 		MetadataController.sanitizedSend(
 			res,
-			await this.handler.fetchMetadata(hash)
+			await this.service.fetchMetadata(hash)
 		);
 	};
 }
