+++ src/controllers/metadata/index.ts
@@ -1,0 +1,1 @@
+export { default as Metadata } from './MetadataController';
