+++ src/controllers/transactions/TransactionsFeeEstimateController.ts
@@ -1,9 +1,8 @@
 import { ApiPromise } from '@polkadot/api';
 
-import ApiHandler from '../ApiHandler';
-import { IPostRequestHandler, ITx } from '../types/request_types';
-import AbstractController from './AbstractController';
-import TxArtifactsController from './TxArtifactsController';
+import { TransactionsFeeEstimateService } from '../../services';
+import { IPostRequestHandler, ITx } from '../../types/request_types';
+import AbstractController from '../AbstractController';
 
 /**
  * POST a serialized transaction and receive a fee estimate.
@@ -31,11 +30,11 @@ import TxArtifactsController from './TxArtifactsController';
  * - `query_info`: https://crates.parity.io/pallet_transaction_payment/struct.Module.html#method.query_info
  * - `compute_fee`: https://crates.parity.io/pallet_transaction_payment/struct.Module.html#method.compute_fee
  */
-export default class TxFeeEstimate extends AbstractController {
-	handler: ApiHandler;
+export default class TransactionsFeeEstimateController extends AbstractController<
+	TransactionsFeeEstimateService
+> {
 	constructor(api: ApiPromise) {
-		super(api, '/tx/fee-estimate');
-		this.handler = new ApiHandler(api);
+		super(api, '/tx/fee-estimate', new TransactionsFeeEstimateService(api));
 		this.initRoutes();
 	}
 
@@ -62,9 +61,9 @@ export default class TxFeeEstimate extends AbstractController {
 
 		const hash = await this.api.rpc.chain.getFinalizedHead();
 
-		TxArtifactsController.sanitizedSend(
+		TransactionsFeeEstimateController.sanitizedSend(
 			res,
-			await this.handler.fetchFeeInformation(hash, tx)
+			await this.service.fetchTransactionFeeEstimate(hash, tx)
 		);
 	};
 }
