+++ src/controllers/transactions/index.ts
@@ -1,0 +1,3 @@
+export { default as TransactionsFeeEstimate } from './TransactionsFeeEstimateController';
+export { default as TransactionsSubmit } from './TransactionsSubmitController';
+export { default as TransactionsMaterial } from './TransactionsMaterialController';
