+++ src/services/blocks/index.ts
@@ -1,0 +1,1 @@
+export { default as BlocksService } from './BlocksService';
