+++ src/services/claims/index.ts
@@ -1,0 +1,1 @@
+export { default as ClaimsService } from './ClaimsService';
