+++ src/services/index.ts
@@ -1,6 +1,6 @@
-export * from './transactions';
 export * from './accounts';
 export * from './blocks';
+export * from './pallets';
+export * from './transactions';
 export * from './claims';
 export * from './metadata';
-export * from './pallets';
