+++ src/services/metadata/MetadataService.ts
@@ -1,0 +1,19 @@
+import { Metadata } from '@polkadot/types';
+import { BlockHash } from '@polkadot/types/interfaces';
+
+import AbstractService from '../AbstractService';
+
+export default class MetadataService extends AbstractService {
+	/**
+	 * Fetch `Metadata` in decoded JSON form.
+	 *
+	 * @param hash `BlockHash` to make call at
+	 */
+	async fetchMetadata(hash: BlockHash): Promise<Metadata> {
+		const api = await this.ensureMeta(hash);
+
+		const metadata = await api.rpc.state.getMetadata(hash);
+
+		return metadata;
+	}
+}
