+++ src/services/metadata/index.ts
@@ -1,0 +1,1 @@
+export { default as MetadataService } from './MetadataService';
