+++ src/services/pallets/index.ts
@@ -1,0 +1,1 @@
+export { default as PalletsStakingProgressService } from './PalletsStakingProgressService';
