+++ src/services/transactions/index.ts
@@ -1,0 +1,3 @@
+export { default as TransactionsSubmitService } from './TransactionsSubmitService';
+export { default as TransactionsFeeEstimateService } from './TransactionsFeeEstimateService';
+export { default as TransactionsMaterialService } from './TransactionsMaterialService';
